Add abstract class structure to Action/Trigger entity classes and refactor
