+++ src/main/java/com/google/flourbot/entity/Macro.java
@@ -3,6 +3,9 @@ package com.google.flourbot.entity;
 import java.util.Map;
 import java.lang.Object;
 
+import com.google.flourbot.entity.action.Action;
+import com.google.flourbot.entity.trigger.Trigger;
+
 public final class Macro {
     private final String creatorId;
     private final String macroName;
