--- src/main/java/com/google/flourbot/entity/Trigger.java
@@ -1,15 +1,0 @@
-package com.google.flourbot.entity;
-
-import java.util.Map;
-
-public final class Trigger {
-    
-    private final String triggerCommand;
-    private final String triggerType;  
-
-    public Trigger (Map<String, Object> document) {
-        this.triggerCommand = (String) document.get("command");
-        this.triggerType = (String) document.get("type");
-    }
-
-}
