+++ src/main/java/com/google/flourbot/entity/action/Action.java
@@ -1,0 +1,9 @@
+package com.google.flourbot.entity.action;
+
+import java.util.Map;
+
+public abstract class Action {
+    public enum ActionType { SHEET_APPEND }
+    public ActionType getActionType();
+}
+
