+++ src/main/java/com/google/flourbot/entity/trigger/Trigger.java
@@ -1,0 +1,9 @@
+package com.google.flourbot.entity.trigger;
+
+import java.util.Map;
+
+public abstract class Trigger {
+    public enum TriggerType { COMMAND_TRIGGER }
+    public TriggerType getTiggerType();
+}
+
